refactor: separate SQLite query port from CLI runtime
