+++ src/infra/adapters/sqlite/sqlite3/parser/command_tag.rs
@@ -62,7 +62,7 @@ fn sqlite_side_effect_tag(query: &str) -> Option<CommandTag> {
     }
 }
 
-pub(in crate::adapters::sqlite::sqlite3) fn sqlite_statement_tags(
+pub(in crate::adapters::sqlite) fn sqlite_statement_tags(
     statements: &[&str],
     changes: &HashMap<usize, usize>,
 ) -> Vec<CommandTag> {
@@ -162,7 +162,7 @@ fn savepoint_frame_index(
         })
 }
 
-pub(in crate::adapters::sqlite::sqlite3) fn aggregate_sqlite_command_tag(
+pub(in crate::adapters::sqlite) fn aggregate_sqlite_command_tag(
     tags: &[CommandTag],
 ) -> Option<CommandTag> {
     let effective = discard_rolled_back(tags);
@@ -178,7 +178,7 @@ pub(in crate::adapters::sqlite::sqlite3) fn aggregate_sqlite_command_tag(
     tags.last().cloned()
 }
 
-pub(in crate::adapters::sqlite::sqlite3) fn command_tag_result(
+pub(in crate::adapters::sqlite) fn command_tag_result(
     query: &str,
     tag: CommandTag,
     elapsed: u64,
@@ -189,9 +189,7 @@ pub(in crate::adapters::sqlite::sqlite3) fn command_tag_result(
         .with_command_tag(tag)
 }
 
-pub(in crate::adapters::sqlite::sqlite3) fn statement_counts_as_select_tag(
-    statement: &str,
-) -> bool {
+pub(in crate::adapters::sqlite) fn statement_counts_as_select_tag(statement: &str) -> bool {
     let keyword = first_keyword(statement);
     keyword.eq_ignore_ascii_case("SELECT") || keyword.eq_ignore_ascii_case("WITH")
 }
