+++ src/infra/adapters/sqlite/sqlite3/parser/mod.rs
@@ -2,16 +2,16 @@ mod command_tag;
 pub(in crate::adapters::sqlite::sqlite3) mod lexer;
 mod output;
 
-pub(super) use command_tag::{
+pub(in crate::adapters::sqlite) use command_tag::{
     aggregate_sqlite_command_tag, command_tag_result, sqlite_statement_tags,
     statement_counts_as_select_tag,
 };
-pub(super) use lexer::{
+pub(in crate::adapters::sqlite) use lexer::{
     SqliteStatementPlan, append_changes_query_for_plan, is_sqlite_rerunnable_export_query,
     reject_sqlite_fsdir, sqlite_adhoc_execution_query_for_plan, sqlite_empty_result_sentinel,
     sqlite_export_not_rerunnable_error, sqlite_probe_marker, sqlite_statement_plan,
 };
-pub(super) use output::{
+pub(in crate::adapters::sqlite) use output::{
     last_sqlite_result_set, parse_affected_rows, parse_count_result, quoted_to_query_result,
     strip_sqlite_probes,
 };
